[Refactor/#115] UI 상태를 model 패키지에서 관리하도록 수정
