+++ feature/submit/src/main/java/com/ilsangtech/ilsang/feature/submit/WordsQuizSubmitViewModel.kt
@@ -8,6 +8,9 @@ import androidx.navigation.toRoute
 import com.ilsangtech.ilsang.core.domain.MissionRepository
 import com.ilsangtech.ilsang.core.domain.QuestRepository
 import com.ilsangtech.ilsang.core.domain.QuizRepository
+import com.ilsangtech.ilsang.feature.submit.model.SubmitQuestUiState
+import com.ilsangtech.ilsang.feature.submit.model.SubmitResultUiState
+import com.ilsangtech.ilsang.feature.submit.model.WordsQuizUiState
 import com.ilsangtech.ilsang.feature.submit.navigation.WordsQuizSubmitRoute
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
